refactor(guard): extract policy source context and unify scope labels
Move remembered-rule enrichment into store_policy_source_context.py and
use scopeLabel policy variant across Policy tabs per review feedback.
